requirements.txt
Django==6.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
requests==2.32.3
psycopg2-binary==2.9.10
python-dotenv==1.0.1
whitenoise==6.7.0
gunicorn==22.0.0
dj-database-url==2.2.0